requirements.txt
web3>=6.0.0
torch>=2.0.0
eth-typing>=3.0.0
eth-account>=0.8.0
aiohttp>=3.8.0
pytest>=7.0.0
solc-select>=1.0.0
python-dotenv>=1.0.0
ipfs-api>=0.8.0
pyyaml>=6.0.0
cryptography>=41.0.0
iota-client>=0.2.0
libp2p>=0.1.5
aioipfs>=0.6.2
transformers>=4.30.0
pillow>=9.5.0
pytesseract>=0.3.10
pdfplumber>=0.9.0
requests>=2.31.0
filecoin-api-client>=0.1.0
arweave-python>=1.0.0
